requirements.txt
Flask==2.1.1
boto3==1.24.5
gitpython==3.1.29
